feat: refactor AnkiParser to streamline data retrieval and enhance database handling
