+++ src/features/AnkiParser/hooks/useAnkiParser.ts
@@ -19,10 +19,7 @@ export const useAnkiParser = (type: string) => {
     const deck = new Deck(type, sqlClient, extractor);
     await deck.init();
 
-    const mediaMap = await deck.getMedia();
-
-    const notesRaw = await deck.getNotes();
-    const models = await deck.getModels();
+    const { mediaMap, models, notesRaw } = await deck.getCollectedData();
 
     const notes = Object.values(notesRaw).map((note) => {
       const model = models[note.mid] as { flds: { name: string }[] };
@@ -67,7 +64,8 @@ export const useAnkiParser = (type: string) => {
       };
       return importData;
     });
-    if (!fastCacheFile) {
+
+    if (fastCacheFile) {
       await add(fastCacheFile!);
     }
