+++ src/features/AnkiParser/lib/constants.ts
@@ -1,7 +1,7 @@
-export const DB_FILES = {
-  MODERN: "collection.anki21b", // modern version
-  LEGACY: "collection.anki21", // legacy
-  OLD: "collection.anki2", // oldest version
+export enum DB_FILES  {
+  MODERN = "collection.anki21b",
+  LEGACY = "collection.anki21",
+  OLD = "collection.anki2",
 };
 
 export const FAST_MEMORY_CLOUD_FILE_NAME = "cloud-url.json";
