+++ src/features/AnkiParser/model/db/Anki21Db.ts
@@ -1,0 +1,3 @@
+import { Db } from "./Db";
+
+export default class Anki21Db extends Db {}
