+++ src/features/AnkiParser/model/db/Anki2Db.ts
@@ -1,0 +1,3 @@
+import { Db } from "./Db";
+
+export default class Anki2Db extends Db {}
